+++ orizon-os-x86_64/kernel/gui/terminal.c
@@ -3822,7 +3822,7 @@ static void term_run_desktop(terminal_t *term, const char *cmd) {
     term_puts_t(term, "  desktop dispatch cyclenext|swapnext|swapwindow|focusmaster|swapwithmaster - Hyprland-like actions\n");
     term_puts_t(term, "  desktop dispatch movefocus <l|r|u|d|next|prev> - Directional tiled focus\n");
     term_puts_t(term, "  desktop dispatch focuswindow <id|0xaddr|class:app|title:text> - Focus a matching client\n");
-    term_puts_t(term, "  desktop dispatch layoutmsg <msg> - orientation/splitratio/masterratio actions\n");
+    term_puts_t(term, "  desktop dispatch layoutmsg <msg> - orientation/splitratio/masterratio/nmaster actions\n");
     term_puts_t(term, "  desktop dispatch resizeactive <x> <y> - Keyboard tiling ratio resize\n");
     term_puts_t(term, "  desktop dispatch submap <name|reset> - Set active submap\n");
     term_puts_t(term, "  desktop reset           - Disable and restore default policy\n");